+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/model/ParameterData.kt
@@ -7,4 +7,13 @@ data class ParameterData(
     val type: TypeData,
     val annotations: List<ParameterAnnotation> = emptyList(),
     val hasRequestBuilderAnno: Boolean = false
-)
+) {
+    inline fun <reified T> findAnnotationOrNull(): T? {
+        return this.annotations.firstOrNull { it is T } as? T
+    }
+
+    inline fun <reified T> hasAnnotation(): Boolean {
+        return this.findAnnotationOrNull<T>() != null
+    }
+
+}