+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/requestData/RequestDataArgumentNode.kt
@@ -1,6 +1,7 @@
 package de.jensklingenberg.ktorfit.requestData
 
-import de.jensklingenberg.ktorfit.hasAnnotation
+
+import de.jensklingenberg.ktorfit.generator.requestDataClass
 import de.jensklingenberg.ktorfit.model.FunctionData
 import de.jensklingenberg.ktorfit.model.ParameterData
 import de.jensklingenberg.ktorfit.model.annotations.Body
@@ -30,6 +31,7 @@ fun getRequestDataArgumentText(functionData: FunctionData): String {
     val body = getBodyDataText(functionData.parameterDataList)
     //URL
     val urlPath = getRelativeUrlArgumentText(functionData.httpMethodAnnotation, functionData.parameterDataList)
+    val pathsText = getPathsText(functionData.parameterDataList)
     val queryText = getQueryArgumentText(functionData.parameterDataList)
     val fieldsText = getFieldArgumentsText(functionData.parameterDataList)
     val partsText = getPartsArgumentText(functionData.parameterDataList)
@@ -46,8 +48,9 @@ fun getRequestDataArgumentText(functionData: FunctionData): String {
         fieldsText,
         partsText,
         builderText,
-        qualifiedTypeName
+        qualifiedTypeName,
+        pathsText
     ).filter { it.isNotEmpty() }.joinToString(",\n") { it }
 
-    return "val requestData = RequestData($args) \n"
+    return "val requestData = ${requestDataClass.name}($args) \n"
 }